+++ apps/earn-protocol/components/layout/VaultOpenView/VaultOpenViewComponent.tsx
@@ -26,10 +26,10 @@ import { VaultOpenHeaderBlock } from '@/components/layout/VaultOpenView/VaultOpe
 import { VaultSimulationGraph } from '@/components/layout/VaultOpenView/VaultSimulationGraph'
 import { TransactionHashPill } from '@/components/molecules/TransactionHashPill/TransactionHashPill'
 import { HistoricalYieldChart } from '@/components/organisms/Charts/HistoricalYieldChart'
-import { RebalancingActivity } from '@/components/organisms/RebalancingActivity/RebalancingActivity'
-import { UserActivity } from '@/components/organisms/UserActivity/UserActivity'
-import { VaultExposure } from '@/components/organisms/VaultExposure/VaultExposure'
 import { TransactionAction } from '@/constants/transaction-actions'
+import { RebalancingActivity } from '@/features/rebalance-activity/components/RebalancingActivity/RebalancingActivity'
+import { UserActivity } from '@/features/user-activity/components/UserActivity/UserActivity'
+import { VaultExposure } from '@/features/vault-exposure/components/VaultExposure/VaultExposure'
 import { useAmount } from '@/hooks/use-amount'
 import { useClient } from '@/hooks/use-client'
 import { usePosition } from '@/hooks/use-position'
