+++ apps/earn-protocol/features/vault-exposure/table/columns.tsx
